date comparison

+++ app/src/main/java/com/example/magnarandroidcodechallenge/viewmodel/MagnarViewModel.kt
@@ -7,6 +7,8 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import com.example.magnarandroidcodechallenge.api.RetrofitInstance
 import com.example.magnarandroidcodechallenge.api.RetrofitService
+import com.google.gson.Gson
+import com.google.gson.reflect.TypeToken
 import com.opencsv.CSVReader
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
@@ -15,11 +17,14 @@ import kotlinx.coroutines.launch
 import okhttp3.ResponseBody
 import retrofit2.Callback
 import java.io.*
+import java.text.SimpleDateFormat
 import java.util.*
+import java.util.concurrent.TimeUnit
 
 
 class MagnarViewModel(
-    private val fileDir: File
+    private val fileDir: File,
+    private val sharedPreferences: SharedPreferences
 ): ViewModel() {
 
     private var dirPath: String = "$fileDir/CSVFiles"
@@ -31,6 +36,8 @@ class MagnarViewModel(
     private val _csvData: MutableLiveData<List<String>> = MutableLiveData()
     val csvData: LiveData<List<String>> get() = _csvData
 
+    private val lastUpdated = Date(sharedPreferences.getLong("lastUpdated", 0))
+    // 16693457
     init {
         val dirFile = File(dirPath)
         if (!dirFile.exists()) {
@@ -41,8 +48,15 @@ class MagnarViewModel(
         csvFileName = File(file)
 
         if(csvFileName.exists()){
-            Log.d("ExistingFile", "File Updated")
-            csvFileName.delete()
+            if(lastUpdated.time + TimeUnit.DAYS.toMillis(1) < Date().time){
+                csvFileName.delete()
+                downloadFile()
+            } else {
+                loadList()
+            }
+        }
+        else{
+            downloadFile()
         }
 
     }
@@ -77,17 +91,17 @@ class MagnarViewModel(
             val outputStream: OutputStream = csvFileName.outputStream()
             inputStream.copyTo(outputStream)
 
+            sharedPreferences.edit().putLong("lastUpdated", Date().time).apply()
             CoroutineScope(Dispatchers.IO).launch {
                 val csvData = async { readFileData() }
                 _csvData.postValue(csvData.await())
             }
-
         } catch (e: IOException) {
             e.printStackTrace()
         }
     }
 
-    private fun readFileData(): List<String> {
+    private suspend fun readFileData(): List<String> {
         val result = mutableListOf<String>()
         try {
             val bufferedReader = BufferedReader(FileReader(csvFileName))
@@ -96,10 +110,28 @@ class MagnarViewModel(
             for (row in data) {
                 result.add(row[0].replace("=", ""))
             }
+
         } catch (e: IOException) {
             e.printStackTrace()
         }
+
         Log.d("File Read", "readFileData: ${result.size}")
+        storeList(result)
+
         return result
+    }
+
+    private fun storeList(list: List<String>){
+        val gson = Gson()
+        val json = gson.toJson(list)
+        sharedPreferences.edit().putString("csvData", json).apply()
+    }
+
+    private fun loadList(){
+        val gson = Gson()
+        val json = sharedPreferences.getString("csvData", null)
+        val type = object : TypeToken<List<String>>() {}.type
+        val list = gson.fromJson<List<String>>(json, type)
+        _csvData.postValue(list)
     }
 }
